Added local test for CreateCustomer request
